create search and update view

+++ routes/web.php
@@ -24,37 +24,34 @@ Route::get('/', function () {
     return view('welcome');
 })->name('welcome');
 
-
 Route::middleware('auth')->group(function () {
     Route::get('/app', [AuthController::class, 'index'])->name('app');
+    Route::get('/permissions', [PermissionController::class, 'index'])->name('permission.index');
     Route::get('/roles', [RoleController::class, 'index'])->name('role.index');
     Route::get('/users', [UserController::class, 'index'])->name('user.index');
-});
 
-Route::middleware('check.permission.role')->group(function () {
-    Route::get('/roles', [RoleController::class, 'index'])->name('role.index');
-    Route::get('/users', [UserController::class, 'index'])->name('user.index');
-    Route::get('role/create', [RoleController::class, 'create'])->name('role.create');
-    Route::get('role/edit/{id}', [RoleController::class, 'edit'])->name('role.edit');
-    Route::get('user/create', [UserController::class, 'create'])->name('user.create');
-    Route::get('user/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
-    Route::delete('/destroy/{id}', [RoleController::class, 'destroy'])->name('role.destroy');
-    Route::delete('/destroy{id}', [UserController::class, 'destroy'])->name('user.destroy');
-});
-//method-post roles
-Route::post('role/store', [RoleController::class, 'store'])->name('role.store');
-Route::post('role/update/{id}', [RoleController::class, 'update'])->name('role.update');
-//method-post user
-Route::post('user/store', [UserController::class, 'store'])->name('user.store');
-Route::post('user/update/{id}', [UserController::class, 'update'])->name('user.update');
+    Route::middleware('check.permission.role')->group(function () {
+        Route::get('permissions/create', [PermissionController::class, 'create'])->name('permission.create');
+        Route::post('permissions/store', [PermissionController::class, 'store'])->name('permission.store');
+        Route::get('permissions/edit/{id}', [PermissionController::class, 'edit'])->name('permission.edit');
+        Route::post('permissions/update/{id}', [PermissionController::class, 'update'])->name('permission.update');
+        Route::delete('permissions/destroy/{id}', [PermissionController::class, 'destroy'])->name('permission.destroy');
+        
+        Route::get('/role/create', [RoleController::class, 'create'])->name('role.create');
+        Route::post('/role/store', [RoleController::class, 'store'])->name('role.store');
+        Route::get('/role/edit/{id}', [RoleController::class, 'edit'])->name('role.edit');
+        Route::post('/role/update/{id}', [RoleController::class, 'update'])->name('role.update');
+        Route::delete('/role/destroy/{id}', [RoleController::class, 'destroy'])->name('role.destroy');
 
-//Permissions
-Route::get('/permissions', [PermissionController::class, 'index'])->name('permission.index');
-Route::get('permissions/create', [PermissionController::class, 'create'])->name('permission.create');
-Route::post('permissions/store', [PermissionController::class, 'store'])->name('permission.store');
-Route::get('permissions/edit/{id}', [PermissionController::class, 'edit'])->name('permission.edit');
-Route::post('permissions/update/{id}', [PermissionController::class, 'update'])->name('permission.update');
-Route::delete('permissions/destroy/{id}', [PermissionController::class, 'destroy'])->name('permission.destroy');
+        Route::get('/user/create', [UserController::class, 'create'])->name('user.create');
+        Route::post('/user/store', [UserController::class, 'store'])->name('user.store');
+        Route::get('/user/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
+        Route::post('/user/update/{id}', [UserController::class, 'update'])->name('user.update');
+        Route::delete('/user/destroy{id}', [UserController::class, 'destroy'])->name('user.destroy');
+
+    });
+
+});
 
 //AUTH
 Route::get('/login', [AuthController::class, 'showLogin'])->name('login');
